Update README.md
We had a bug where in-app route changes weren't firing when adding the script dynamically. Setting the tracking ID in the module configuration as described resolved this, but it wasn't easy to find. I don't think it's uncommon to set the tracking ID based on the environment this way, so I thinks it's worth documenting this.